+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Make resource callable, rename `#respond` to `#callable` (#3)
+
 ## 0.2 2022-01-13
 
 - Implement `FileTree` (#2)